+++ src/typechecker.js
@@ -15,6 +15,7 @@ var _       = require("lodash");
 var ts      = require("typescript");
 
 var TypecheckerSymbols = require("./model/OJSymbolTyper").TypecheckerSymbols;
+var Location           = require("./model/OJSymbolTyper").Location;
 
 
 var sBlacklistCodes  = [ 2417 ];
@@ -32,18 +33,22 @@ var sReasonTemplates = [
     { code: 2339, type: "-", text: "No known instance method: -[{1} {0}]" },
 
     // 2415: Class '{0}' incorrectly extends base class '{1}'.
-    // Types of property 'buildElements' are incompatible.
+    // 2326: Types of property 'buildElements' are incompatible.
     { code: 2415, text: "'{2415:0}' and '{2415:1}' have incompatible method '{2326:0}'" },
 
     // 2420: Class '{0}' incorrectly implements interface '{1}'.
     // 2324: Property '{0}' is missing in type '{1}'.
-    { code: 2420, text: "Method '{2324:0}' in protocol '{2420:1}' not implemented" }
+    { code: 2420, next: 2324, text: "Method '{2324:0}' in protocol '{2420:1}' not implemented" },
+
+    // 2420: Class '{0}' incorrectly implements interface '{1}'.
+    // 2326: Types of property 'buildElements' are incompatible.
+    { code: 2420, next: 2326, text: "'{2420:0}' and protocol '{2420:1}' have incompatible method '{2326:0}'" }
 ];
 
 
 var sReasonTemplateMap;
 
-function getReasonTemplate(code, sawClass, sawProtocol, sawMethod, sawStatic) {
+function getReasonTemplate(code, nextCode, sawClass, sawProtocol, sawMethod, sawStatic) {
     var types = [ ];
 
     if (sawMethod)   types.push(sawStatic ? "+" : "-");
@@ -57,12 +62,12 @@ function getReasonTemplate(code, sawClass, sawProtocol, sawMethod, sawStatic) {
         sReasonTemplateMap = { };
 
         _.each(sReasonTemplates, function(t) {
-            sReasonTemplateMap["" + t.code + (t.type || "")] = t;
+            sReasonTemplateMap["" + t.code + "." + (t.next || 0) + (t.type || "")] = t;
         });
     }
 
     for (var i = 0, length = types.length; i < length; i++) {
-        var reason = sReasonTemplateMap["" + code + (types[i] || "")];
+        var reason = sReasonTemplateMap["" + code + "." + nextCode + (types[i] || "")];
         if (reason) return reason;
     }
 
@@ -73,18 +78,30 @@ function getReasonTemplate(code, sawClass, sawProtocol, sawMethod, sawStatic) {
 /*
     See https://github.com/Microsoft/TypeScript/wiki/Using-the-Compiler-API
 */
-var sLibrarySource;
+var sCachedLibFileContents = { };
 
-function ts_transform(defs, code, options)
+function ts_transform(defs, code, options, libFileName)
 {
+    var libFileContents = sCachedLibFileContents[libFileName];
+
     // Eventually, there should be an option/way to switch to "lib.core.d.ts"
     // Keep in sync with: https://github.com/Microsoft/TypeScript/issues/494 ?
     //
-    if (!sLibrarySource) {
-        sLibrarySource = fs.readFileSync(path.join(path.dirname(require.resolve('typescript')), 'lib.d.ts')).toString();
+    if (!libFileContents) {
+        try {
+            libFileContents = fs.readFileSync(path.join(path.dirname(require.resolve("typescript")), libFileName)).toString();
+            sCachedLibFileContents[libFileName] = libFileContents;
+        } catch (e) {
+            libFileName = null;
+            libFileContents = " ";
+        }
     }
 
-    var contentMap = { "defs.ts": defs, "lib.d.ts": sLibrarySource, "code.ts": code, };
+    var contentMap = {
+        "defs.ts": defs,
+        "code.ts": code,
+        "lib.d.ts": libFileContents
+    };
 
     var program = ts.createProgram(_.keys(contentMap), options || { }, {
         getSourceFile: function(filename, languageVersion) {
@@ -107,7 +124,7 @@ function ts_transform(defs, code, options)
 }
 
 
-function TypeChecker(model, generator, files, noImplicitAny)
+function TypeChecker(files, model, options)
 {
     var defsResult = this._generateDefs(model);
     this._defs = defsResult.defs;
@@ -122,7 +139,7 @@ function TypeChecker(model, generator, files, noImplicitAny)
 
     this._files = files;
     this._model = model;
-    this._noImplicitAny = noImplicitAny;
+    this._options = options;
 }
 
 
@@ -135,25 +152,13 @@ TypeChecker.prototype._generateDefs = function(model)
         var parameters = [ ];
 
         for (var i = 0, length = method.parameterTypes.length; i < length; i++) {
-            var parameterType = method.parameterTypes[i];
             var variableName  = method.variableNames[i] || ("a" + i);
-
-            if (parameterType == "id") {
-                parameterType = TypecheckerSymbols.IdUnion;
-            } else {
-                parameterType = symbolTyper.toTypecheckerType(parameterType);
-            }
+            var parameterType = symbolTyper.toTypecheckerType(method.parameterTypes[i], Location.DeclarationParameter);
 
             parameters.push(variableName + " : " + parameterType);
         }
 
-        var returnType;
-
-        if (method.returnType == "id") {
-            returnType = TypecheckerSymbols.IdIntersection;
-        } else {
-            returnType = symbolTyper.toTypecheckerType(method.returnType, ojClass);
-        }
+        var returnType = symbolTyper.toTypecheckerType(method.returnType, Location.DeclarationReturn, ojClass);
 
         return methodName + (method.optional ? "?" : "") + "(" + parameters.join(", ") + ") : " + returnType + ";";
     }
@@ -304,7 +309,7 @@ TypeChecker.prototype._generateDefs = function(model)
     var classSymbols  = [ ];
     var staticSymbols = [ ];
 
-    var runtimeContents = fs.readFileSync(dirname(__filename) + "/runtime.d.ts") + "\n";
+    var runtimeContents = fs.readFileSync(dirname(__filename) + "/../runtime/runtime.d.ts") + "\n";
     var lines = runtimeContents.split("\n");
     var lineMap = [ ];
 
@@ -347,12 +352,25 @@ TypeChecker.prototype._generateDefs = function(model)
         });
     });
 
+    _.each(model.structs, function(ojStruct) {
+        var structSymbol = symbolTyper.getSymbolForStructName(ojStruct.name);
+
+        lines.push("interface " + structSymbol + "{");
+
+        _.each(ojStruct.variables, function(variable) {
+            lines.push(variable.name + " : " + symbolTyper.toTypecheckerType(variable.annotation));
+        });
+
+        lines.push("}");
+    });
+
     lines.push("declare class " + TypecheckerSymbols.GlobalType + " {");
     _.each(model.globals, function(ojGlobal) {
+        var name       = symbolTyper.getSymbolForIdentifierName(ojGlobal.name);
         var annotation = _.clone(ojGlobal.annotation);
 
         if (_.isArray(annotation)) {
-            var line = ojGlobal.name;
+            var line = name;
             var returnType = annotation.shift();
 
             line += "(" + _.map(annotation, function(a, index) {
@@ -364,7 +382,7 @@ TypeChecker.prototype._generateDefs = function(model)
             lines.push(line);
 
         } else {
-            lines.push(ojGlobal.name + " : " + symbolTyper.toTypecheckerType(annotation) + ";");
+            lines.push(name + " : " + symbolTyper.toTypecheckerType(annotation) + ";");
         }
     });
     lines.push("}");
@@ -392,7 +410,7 @@ TypeChecker.prototype.check = function(callback)
     var codeLineMap   = this._codeLineMap;
 
     var symbolTyper   = this._model.getSymbolTyper();
-    var noImplicitAny = this._noImplicitAny;
+    var options       = this._options;
 
     var prependLineCount = this._prependLineCount;
 
@@ -419,8 +437,6 @@ TypeChecker.prototype.check = function(callback)
                 // This doesn't account for prepended files (added by the modifier), so
                 // offset accordingly
 
-                console.log(defsLine, entry.start, entry.end, entry.mapped, prependLineCount);
-
                 return getFileLineWithCodeLine(entry.mapped + prependLineCount);
             }
         }
@@ -429,16 +445,18 @@ TypeChecker.prototype.check = function(callback)
     }
 
     function makeHintsWithDiagnostic(diagnostic) {
-        var lineColumn  = diagnostic.file.getLineAndCharacterOfPosition(diagnostic.start);
+        var lineColumn  = diagnostic.file ? diagnostic.file.getLineAndCharacterOfPosition(diagnostic.start) : { line: 0, column: 0 };
+        var fileName    = diagnostic.file ? diagnostic.file.fileName : "";
         var fileLine    = null;
 
-        if (diagnostic.file.fileName == "code.ts") {
+        if (fileName == "code.ts") {
             fileLine = getFileLineWithCodeLine(lineColumn.line);
-        } else if (diagnostic.file.fileName == "defs.ts") {
+        } else if (fileName == "defs.ts") {
             fileLine = getFileLineWithDefsLine(lineColumn.line);
         }
 
         var code        = diagnostic.code;
+        var next;
         var quotedMap   = { };
         var sawStatic   = false;
         var sawMethod   = false;
@@ -481,8 +499,10 @@ TypeChecker.prototype.check = function(callback)
 
                     a1 = symbolTyper.fromTypecheckerType(a1);
 
-                    quotedMap[i] = a1;
-                    quotedMap["" + code + ":" + i] = a1;
+                    var key = "" + code + ":" + i;
+
+                    if (!quotedMap[i])   quotedMap[i]   = a1;
+                    if (!quotedMap[key]) quotedMap[key] = a1;
 
                     i++;
 
@@ -494,13 +514,14 @@ TypeChecker.prototype.check = function(callback)
             while (messageText) {
                 parseMessageText(messageText);
                 messageText = messageText.next;
+                if (!next) next = messageText ? messageText.code : 0;
             }
         }());
 
         // Now look up the friendlier reason string from the map
         //
         (function() {
-            var template = getReasonTemplate(code, sawClass, sawProtocol, sawMethod, sawStatic);
+            var template = getReasonTemplate(code, next, sawClass, sawProtocol, sawMethod, sawStatic);
 
             var valid  = true;
             var result = null;
@@ -552,10 +573,12 @@ TypeChecker.prototype.check = function(callback)
         return null;
     }
 
-    var options = { };
-    if (noImplicitAny) options.noImplicitAny = true;
+    var tsOptions = { };
+    if (options["no-implicit-any"]) tsOptions.noImplicitAny = true;
 
-    var hints = _.flatten(_.map(ts_transform(defs, code, options), function(diagnostic) {
+    var tsLibName = options["typescript-lib"] || "lib.d.ts";
+
+    var hints = _.flatten(_.map(ts_transform(defs, code, tsOptions, tsLibName), function(diagnostic) {
         return makeHintsWithDiagnostic(diagnostic);
     }));