+++ Changelog.md
@@ -5,9 +5,10 @@ CHANGELOG
 1.2.0
 - Update to Esprima 2.7 (#38)
 - Added @CLASS, @SEL, @FUNCTION for debugging (#58)
+- @global functions (#68)
 - Category (#55) and protocol (#57) support
 - Throw errors when the compiler will generate invalid code (#60)
-- Major changes to the typechecker (#64, #65)
+- Major changes to the typechecker (#64, #65, #71)
 
 1.1.1 - 1.1.4
 - Minor bug fixes and internal changes